# phase-4-kubernetes/specs/001-phase-2-fullstack/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Task Management API
  description: API for task management with user authentication
  version: 1.0.0
  contact:
    name: API Support
    email: support@taskmanager.example.com

servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.taskmanager.example.com
    description: Production server

paths:
  /api/auth/register:
    post:
      summary: Register a new user
      description: Creates a new user account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/login:
    post:
      summary: Authenticate user
      description: Authenticates user and returns JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "access_token=jwt_token_here; HttpOnly; Secure; SameSite=Lax"
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/{user_id}/tasks:
    get:
      summary: Get all tasks for a user
      description: Returns all tasks associated with the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user whose tasks to retrieve
      responses:
        '200':
          description: Successfully retrieved tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '401':
          description: Unauthorized
        '404':
          description: User not found or unauthorized access
      security:
        - BearerAuth: []
    post:
      summary: Create a new task
      description: Creates a new task for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user creating the task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '404':
          description: User not found or unauthorized access
      security:
        - BearerAuth: []

  /api/{user_id}/tasks/{task_id}:
    get:
      summary: Get a specific task
      description: Returns a specific task for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the task to retrieve
      responses:
        '200':
          description: Successfully retrieved task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Unauthorized
        '404':
          description: Task not found or unauthorized access
      security:
        - BearerAuth: []
    put:
      summary: Update a task
      description: Updates an existing task for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the task to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '404':
          description: Task not found or unauthorized access
      security:
        - BearerAuth: []
    delete:
      summary: Delete a task
      description: Deletes a specific task for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the task to delete
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Task not found or unauthorized access
      security:
        - BearerAuth: []

  /api/{user_id}/tasks/{task_id}/complete:
    patch:
      summary: Toggle task completion
      description: Toggles the completion status of a task for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the task to toggle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskToggleCompletionRequest'
      responses:
        '200':
          description: Task completion status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '404':
          description: Task not found or unauthorized access
      security:
        - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        password:
          type: string
          minLength: 8
          description: User's password (min 8 characters)
          example: securepassword123

    RegisterResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: The created user's ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          description: The user's email
          example: user@example.com
        message:
          type: string
          description: Success message
          example: "Account created successfully"

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        password:
          type: string
          description: User's password
          example: securepassword123

    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          enum: [bearer]
          description: Token type
          default: "bearer"
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
              description: User's ID
              example: "550e8400-e29b-41d4-a716-446655440000"
            email:
              type: string
              format: email
              description: User's email
              example: user@example.com

    TaskCreateRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title (1-200 characters)
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          description: Task description (optional, max 1000 characters)
          example: "Need to buy milk, bread, and eggs"

    TaskUpdateRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Updated task title (1-200 characters)
          example: "Buy groceries and cook dinner"
        description:
          type: string
          maxLength: 1000
          description: Updated task description (max 1000 characters)
          example: "Need to buy milk, bread, eggs, and vegetables"
        completed:
          type: boolean
          description: Updated completion status
          example: true

    TaskToggleCompletionRequest:
      type: object
      required:
        - completed
      properties:
        completed:
          type: boolean
          description: New completion status
          example: true

    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          description: Task ID
          example: 1
        user_id:
          type: string
          format: uuid
          description: ID of the user who owns the task
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          description: Task title
          example: "Buy groceries"
        description:
          type: string
          description: Task description
          example: "Need to buy milk, bread, and eggs"
        completed:
          type: boolean
          description: Whether the task is completed
          example: false
        created_at:
          type: string
          format: date-time
          description: When the task was created
          example: "2026-01-28T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: When the task was last updated
          example: "2026-01-28T10:00:00Z"

    TaskListResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'

    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Error message
          example: "Invalid email format"

security:
  - BearerAuth: []